ParseFloat to Int

+++ ai.js
@@ -375,14 +375,14 @@ function setSnaffleWillGoal () {
     let newVelocity = {x: snaffle.vx, y: snaffle.vy};
     let newPos = {x: snaffle.x, y: snaffle.y};
 
-    if (lineIntersect(snaffle.x, snaffle.y, parseFloat(newPos.x) + parseFloat(newVelocity.x), parseFloat(newPos.y) + parseFloat(newVelocity.y),
+    if (lineIntersect(snaffle.x, snaffle.y, parseInt(newPos.x) + parseInt(newVelocity.x), parseInt(newPos.y) + parseInt(newVelocity.y),
                           goalToProtect.point1.x, goalToProtect.point1.y, goalToProtect.point2.x, goalToProtect.point2.y)) {
       debug('Snaffle ' + snaffle.id + ' too fast, cant be stopped with petri');
       return;
     }
 
     while (Math.abs(newVelocity.x) + Math.abs(newVelocity.y) > 100) {
-      newPos = {x: parseFloat(newPos.x) + parseFloat(newVelocity.x), y: parseFloat(newPos.y) + parseFloat(newVelocity.y)};
+      newPos = {x: parseInt(newPos.x) + parseInt(newVelocity.x), y: parseInt(newPos.y) + parseInt(newVelocity.y)};
       newVelocity = {x: round(newVelocity.x * friction.snaffle), y: round(newVelocity.y * friction.snaffle)};
     }
